Initlize: Init the folder structure for v1.2.2
