--- src/main/kotlin/com/github/srilakshmikanthanp/quicknote/appconsts/AppConsts.kt
@@ -1,62 +1,0 @@
-package com.github.srilakshmikanthanp.quicknote.appconsts
-
-
-/**
- * A Singleton instance that Holds the Constants of the Application
- */
-object AppConsts {
-    /**
-     * A Constant for the Home Page of Application
-     */
-    const val APP_HOME = "https://github.com/srilakshmikanthanp/QuickNote"
-
-    /**
-     * Applicatio Resource Port
-     */
-    const val APP_PORT = 57428
-
-    /**
-     * A Constant for the Min Width of Editor
-     */
-    const val MIN_WIDTH = 100.0
-
-    /**
-     * A Constant for Default Width
-     */
-    const val DEF_WIDTH = 350.0
-
-    /**
-     * A Constant for the Max Width of Editor
-     */
-    const val MAX_WIDTH = 1000.0
-
-    /**
-     * A Constant for the Min Height of Editor
-     */
-    const val MIN_HEIGHT = 100.0
-
-    /**
-     * A Constant for Default Height
-     */
-    const val DEF_HEIGHT = 350.0
-
-    /**
-     * A Constant for the Max Height of Editor
-     */
-    const val MAX_HEIGHT = 550.0
-
-    /**
-     * A Constant for the dark theme
-     */
-    const val DEF_THEME = false
-
-    /**
-     * A Default constant for lock
-     */
-    const val DEF_LOCK = false
-
-    /**
-     * A Constant for the Dfault Text
-     */
-    const val DEF_TEXT = ""
-}
