+++ src/main/kotlin/com/github/srilakshmikanthanp/quicknote/components/NoteEditor.kt
@@ -1,0 +1,1 @@
+package com.github.srilakshmikanthanp.quicknote.editor
