+++ src/main/kotlin/com/github/srilakshmikanthanp/quicknote/settings/AppSetting.kt
@@ -1,0 +1,2 @@
+package com.github.srilakshmikanthanp.quicknote.settings
+
