+++ src/main/kotlin/com/github/srilakshmikanthanp/quicknote/system/NoteIcon.kt
@@ -1,0 +1,1 @@
+package com.github.srilakshmikanthanp.quicknote.system
