+++ src/main/kotlin/com/github/srilakshmikanthanp/quicknote/types/Positions.kt
@@ -1,0 +1,5 @@
+package com.github.srilakshmikanthanp.quicknote.types
+
+enum class Positions {
+    TOP, BOTTOM, LEFT, RIGHT, NORMAL
+}
