+++ src/main/kotlin/com/github/srilakshmikanthanp/quicknote/utility/Functions.kt
@@ -1,0 +1,1 @@
+package com.github.srilakshmikanthanp.quicknote.utility
