+++ src/main/kotlin/module-info.java
@@ -1,8 +1,8 @@
 module quicknote {
+    exports com.github.srilakshmikanthanp.quicknote;
+
+    requires javafx.controls;
     requires java.prefs;
     requires java.desktop;
-    requires javafx.controls;
     requires kotlin.stdlib;
-
-    exports com.github.srilakshmikanthanp.quicknote;
 }
