+++ src/main/resources/quicknote.properties
@@ -1,2 +1,2 @@
-version=${project.version}
 artifactId=${project.artifactId}
+version=${project.version}
