--- src/main/resources/styles/Dark.css
@@ -1,6 +1,0 @@
-@import url('./Both.css');
-
-* {
-    -fx-bg-color: #121212;
-    -fx-tumb-color: #363535;
-}
