--- src/main/resources/styles/Light.css
@@ -1,6 +1,0 @@
-@import url('./Both.css');
-
-* {
-    -fx-bg-color: white;
-    -fx-tumb-color: #cfcccc;
-}
